remove unused pj_context code, move error util
