+++ src/proj-utils.h
@@ -3,3 +3,4 @@
 #include <proj.h>
 
 const char* proj_type_name(PJ_TYPE type);
+void stop_proj_error(PJ_CONTEXT* ctx);
